refactor: - legacy websocket removed.
